redirect_service/requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
cassandra-driver==3.29.3
